dev-requirements.txt

# Testing
pytest
pytest-httpx
trio
pytest-asyncio
